+++ src/constants/index.js
@@ -1,4 +1,4 @@
-const VERSION = '1.5.2'
+const VERSION = '1.6.0'
 const BLOCK_ROWS = 500
 const CLUSTER_BLOCKS = 4